+++ include/server/api/base_parameters_grammar.hpp
@@ -78,10 +78,17 @@ struct BaseParametersGrammar : boost::spirit::qi::grammar<Iterator, Signature>
         : BaseParametersGrammar::base_type(root_rule)
     {
         const auto add_hint = [](engine::api::BaseParameters &base_parameters,
-                                 const boost::optional<std::string> &hint_string) {
-            if (hint_string)
+                                 const std::vector<std::string> &hint_strings) {
+            if (!hint_strings.empty())
             {
-                base_parameters.hints.emplace_back(engine::Hint::FromBase64(hint_string.get()));
+                std::vector<engine::SegmentHint> location_hints(hint_strings.size());
+                std::transform(hint_strings.begin(),
+                               hint_strings.end(),
+                               location_hints.begin(),
+                               [](const auto &hint_string) {
+                                   return engine::SegmentHint::FromBase64(hint_string);
+                               });
+                base_parameters.hints.push_back(engine::Hint{std::move(location_hints)});
             }
             else
             {
@@ -145,10 +152,11 @@ struct BaseParametersGrammar : boost::spirit::qi::grammar<Iterator, Signature>
                         (-(qi::double_ | unlimited_rule) %
                          ';')[ph::bind(&engine::api::BaseParameters::radiuses, qi::_r1) = qi::_1];
 
-        hints_rule = qi::lit("hints=") >
-                     (-qi::as_string[qi::repeat(engine::ENCODED_HINT_SIZE)[base64_char]])[ph::bind(
-                         add_hint, qi::_r1, qi::_1)] %
-                         ';';
+        hints_rule =
+            qi::lit("hints=") >
+            (*qi::as_string[qi::repeat(engine::ENCODED_SEGMENT_HINT_SIZE)[base64_char]])[ph::bind(
+                add_hint, qi::_r1, qi::_1)] %
+                ';';
 
         generate_hints_rule =
             qi::lit("generate_hints=") >